feature: update the deployment method
now deploying to Google Cloud Run can simply be deployed with the
deploy.sh script. Users can just run the command `sh deploy.sh` on the
terminal then the deployment will automatically proceed. There are
things that must be set up initially to proceed with the
auto-deployment.

the following environment variables must be set on the .env.local and
.env.local should not be shared:

- IMAGE_NAME
- IMAGE_TAG
- GCP_PROJECT_ID
- DEPLOY_REGION
- VPC_CONNECTOR (optional)
- VPC_REGION (optional)
- SECRET_ENVIRONMENT_VARIABLE_* (optional: name of the environment variable as a secret on Google Cloud
Run)
- SECRET_NAME_* (optional: name of secret from Google Cloud Secret)

to use the optional Environment Variables such as VPC_CONNECTOR, VPC, or
SECRET, they must be set on the Google Cloud RUN.

note: you must enter the service name at the end of the deployment
process. Put the same service you have done deployment before or put a
new one to create a new service.

Core changes:
- feature: update the HOSTNAME from the environment variable.
- feature: add deploy.sh to automatically run the deployment to the Google Cloud Run.

Misc changes:
- fix: fix the typo. feature: fix the casing on NODE_ENV of environmental.
variable.
- feature: split the securityHeaders from the next.config.js.
