+++ .env
@@ -5,7 +5,7 @@
 #
 # HOST
 # use .env.development.local to override the .env
-HOSTNAME=tp-next-todos-app01-2mjfsfem7q-uc.a.run.app
+HOSTNAME=todo.atalas.dev
 HOST=https://$HOSTNAME:$PORT
 NEXT_PUBLIC_HOST=$HOST
 # API
